implemented search algorithm
implemented search algorithm for the shortest route between a cluster of points connected to each other. 50€ for the guy that reads the changes anyways and can explain what the heck I was doing.

+++ Wegbeschreibung/src/database/Database.java
@@ -6,7 +6,7 @@ public abstract interface Database
 {
     public List<String> getNeighboursOfRoom(String roomName);
 
-    public List<Integer> getCoordinatesOfRoom(String roomName);
+    public List<Double> getCoordinatesOfRoom(String roomName);
 
     public List<String> getRoomsForFloor(String floorName);
